% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xmlExample.R
\name{addImportDef}
\alias{addImportDef}
\title{Manage Import Definitions}
\usage{
addImportDef(node, impdefName, rImportCommand, importParameters = NULL)
}
\arguments{
\item{node}{An ICF node to which the Import Definition will be added}

\item{impdefName}{A string the uniquely identified the Import Definition}

\item{rImportCommand}{The name (as a string, without parenthesis at the end)
of the function to run to import a datafile.  This function should have a
parameter called "filename" and must return a dataframe.}

\item{importParameters}{A named list of parameters OTHER THAN THE \code{filename} PARAMETER to be passed to}
}
\description{
Import definitions are added to an Import Control File (ICF) to store the
description of a command that can be used to create a dataframe from a raw
data file.
}
\details{
Importing data from a file into R presumes knowledge of the file format.  Yet
so long as datafiles share the same format, they are be imported with the
same function.  Import Definition are, in essence, descriptions of a function
that can be used to import a datafile of a particular format.  Once
specified, the Import Definition can be referenced by any file specification
in the ICF (see \code{addDatafile()} as a means of creating file
specifications in an ICF)
}

