% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpUtils.R
\name{calculatedDatumColumns}
\alias{calculatedDatumColumns}
\title{Auto-generated columns for any dpBatch in a dpSet}
\usage{
calculatedDatumColumns(set)
}
\arguments{
\item{set}{A \code{dpSet} object or the path to a \code{dpSet} directory on
disk.}
}
\value{
Returns a character vector of the columns that are automatically
  created by \code{datapuppy} for any given \code{dpSet}.  These columns
  appear in one or more tables in the database associated with the
  \code{dpSet}, but should not (and can not) be specified as a column name in
  any data.frame assciated with a \code{dpBatch}.
}
\description{
For all \code{dpBatches} in a \code{dpSet}, \code{Datapuppy} automatically
generates some columns prior to loading the \code{dpBatchs}'s data into the
database associated with the \code{dpSet}.
}

