% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xmlExample.R
\name{createICFNode}
\alias{createICFNode}
\title{Create Import Control File as an XML node.}
\usage{
createICFNode(path)
}
\arguments{
\item{path}{An optional root path that will be appended to datafile names in the input control file.}
}
\value{
An XMLNode object (see XML package documentation) that represents the skeleton of an empty input control file.
}
\description{
Create Import Control File as an XML node.
}

