% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dpBatch.R
\name{dpBatch}
\alias{dpBatch}
\title{Create and manage batches}
\usage{
dpBatch(batchRecord, setPath, dataValues)
}
\arguments{
\item{batchRecord}{A named \code{list} of values describing the \code{batch}.
Values in the \code{batchRecord} are stored as a record in the batches
table.  \code{Names} attribute of \code{batchRecord} must contain the name
of the database table column where each value is to be stored.}

\item{dataValues}{A \code{data.frame} with the values that are to be loaded
into the database.}

\item{set}{A \code{\link{dpSet}} object describing \code{set} to which the
batch will be added.}
}
\description{
These are details.
}

