% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpLoad.R
\name{dpCommitBatchData}
\alias{dpCommitBatchData}
\title{Commit (upload) data from a dpBatch to the associated database.}
\usage{
dpCommitBatchData(batch)
}
\arguments{
\item{batch}{a \code{\link{dpBatch}} object.}

\item{connection}{An connection object, ideally returned from
\code{dpConnect()} (see also \code{\link{odbcConnect}} from the RODBC
package.)}
}
\description{
Transfers the \code{data.frame} associated with a
  \code{\link{dpBatch}} object into the database described by
  \code{dpBatch}'s \code{\link{dpSet}}.  Any \code{\link{dpTweak}}s that are
  associated with the \code{dpBatch} are processed prior to loading the data
  into the database.
}

