% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R
\name{dpConnect}
\alias{dpConnect}
\title{Connect to a database using a dpConnectionArgs object.}
\usage{
dpConnect(dpConnectionArgs)
}
\arguments{
\item{dpConnectionArgs}{A \code{dpConnectionArgs} object created by the
\code{\link{dpConnectionArgs}} function.}
}
\value{
An DBI connection object (see \code{\link{dbConnect}} in the
  \code{DBI} package)
}
\description{
Open connections to ODBC databases using a \code{dpConnectionArgs} object
}

