% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R
\name{dpConnectionArgs}
\alias{dpConnectionArgs}
\title{Create a dpConnectionArgs object}
\usage{
dpConnectionArgs(odbcConnectFunctionName, arguments)
}
\arguments{
\item{odbcConnectFunctionName}{A \code{character} variable specifying the
function (from the RODBC package) used to connect to a database.}

\item{arguments}{A named \code{list} of arguments passed to
\code{odbcConnectFunction} where the names in the \code{list} are the names
of the function arguments.}
}
\value{
A \code{dpConnectionArgs} object that can be passed to
  \code{dpSet()} or \code{dpConnect()}.
}
\description{
A dpConnectionArgs object is passed to the \code{dpSet()} function to
create a dpSet object. pdConnectionArgs objects specify the connect
function (from the RODBC package) and the arguments passed to that function
in order to create to the set's database.
}

