% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dpSet.R
\name{dpLoadSetParamList}
\alias{dpLoadSetParamList}
\title{When \code{dpSet()} is called, the function creates a file called
"dpSet.rData" in the location specified by the \code{setPath} parameter. That
file contains a list of the parameters passed to \code{dpSet()}.  When a set
is reloaded from disk, the list of the parameters is loaded and passed to
\code{dpSet()} to recreate the set.}
\usage{
dpLoadSetParamList(setPath, validate = TRUE)
}
\description{
\code{dpLoadSetParamList()} returns the list of parameters that was originally
passed to \code{dpSet()} to create the set. By default, the list will be
validated against \code{formals(dpSet)} and an error will be thown if the
call list doesn't match the dpSet function.  If validate=FALSE, the function
will return the loaded parameter list, so that an invalid call list can be
inspected or repaired.
}

