% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dpSet.R
\name{dpSet}
\alias{dpCheckSet}
\alias{dpLoadSet}
\alias{dpSet}
\title{Create and manage datapuppy \code{sets} and \code{batches}}
\usage{
dpSet(setPath, connectionParams, batchRowColumnName, datumValueColumnName,
  datumTypeColumnName, batchNameColumnName, batchesTableName, dataTableName,
  typesTableName)

dpLoadSet(setPath)

dpCheckSet(setPath, reportError = TRUE)
}
\arguments{
\item{setPath}{A character string containing the path to the folder that will
house the \code{set}.  The folder must exist and be empty when a new
\code{set} is created.  The path can be fully specified or relative to the
R working directory.}

\item{batchRowColumnName}{A character string containing the name of a
column in the \code{dataTable} of the database; the column is used to store
the row number of the \code{dataValues data.frame} that was the source of
the datum.  (See \code{\link{dpBatch}} for more information on the
\code{dataValues data.frame}).}

\item{batchNameColumnName}{A character string containing the name of a column
in the \code{batchTable} of the database that stores a unique name for the
batch.  This name should be meaningful to a human to identify the batch,
not the serial number (primary key) assigned to the batch automatically by
datapuppy.}

\item{batchesTableName}{A character string containing the name of the
\code{batchesTable} in the database.}

\item{dataTableName}{A character string containing the name of the
\code{dataTable} in the database.}

\item{typesTableName}{A character string containing the name of the
\code{typesTable} in the database.}

\item{reportError}{Boolean value when set to \code{TRUE} causes
\code{dpCheckSet} to print a message and stops execution.  Set to
\code{FALSE}, any errors are returned in a list (see code{Value} section,
below.)}

\item{conndectionParams}{A \code{dpConnectionParams} object returned from
\code{\link{dpConnectionParams()}} describing the location and credentials
for the database associated with a \code{set}.}
}
\value{
\code{dpSet()} returns a dpSet object that references both the
  storage location (file folder) of the \code{set} and the database
  associated with the \code{set}.  This object should be assigned to a
  variable so that it can be passed to datapuppy functions that help manage
  \code{batches} within the \code{set}.  Note that \code{dbSet()} also saves
  a copy of the parameters passed to \code{dpSet()} (technically, it saves the
  results of \code{as.list(match.call())}) to a file called "dpSet.rData" in
  the folder specified by \code{setPath}.

\code{dpCheckSet()} returns a list containing two character vectors.
  The first vector, named "missingParamters", contains the names of
  parameters required by dpSet() that are missing from the stored set.  The
  second vector, named "extraParamters" contains the names of any extra
  parameters in the stored set, which should be removed.  When
  \code{reportError} is TRUE, the same information is printed as an error
  message to the user, and the list containing "missingParameters" and
  "extraParamters" is returned invisibly (see \code{\link{invisible}}).
}
\description{
Datapuppy \code{sets} are collections of data \code{batches} to be loaded
into a single database.  \code{Batches} are collections of records derived
from a single data file (e.g., a data file downloaded from a data logger or a
spreadsheet that contains field observations).
}
\details{
\code{Sets} are collections of data \code{batches}.  The concept of a
\code{set} was developed assuming that a \code{set} would represent all of
the data stored in a particular database.  Thus, each \code{set} is always
associated with a single database.  (Technically, however, there is no reason
a single database couldn't be associated with and contain multiple
\code{sets}.)

A \code{batch} is a collection of data points that have been imported from a
single data file.  For instance, when a data logger is downloaded, it creates
a data file with many records, where each record may contain observations of
several different metrics. All (or a subset) of those observations can be
collected into a \code{batch}.

Note that datapuppy makes a few assumptions about the columns in the
\code{batchesTable}, \code{dataTable}, and \code{typesTable} of the database:

1) Each table must contain an autonumber field that is designated as the
primary key for the table

2) the \code{dataTable} must contain at least two foreign keys, one that
refers to and is named the same as the primary key column of the
\code{batchesTable}, and one that refers to and is named the same as the
primary key column of the \code{typesTable}.

\code{dpCheckSet} reads the set stored at \code{setPath} and compares the
parameters used to create the stored set the the parameters required by
\code{dpSet()}.
}

