% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xmlExample.R
\name{readICFNode}
\alias{readICFNode}
\title{Read an input control file for processing}
\usage{
readICFNode(file)
}
\arguments{
\item{file}{Name of the import control file to read.}
}
\value{
The contents of the import control file as an XMLNode object (see XML package documentation)
}
\description{
Read an input control file for processing
}

