% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dpUtils.R
\name{unpackDots}
\alias{unpackDots}
\title{examine ... parameter and creates variables in calling environment.}
\usage{
unpackDots(...)
}
\description{
examine ... parameter and creates variables in calling environment.
}

