% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dpUtils.R
\name{dpCheckPath}
\alias{dpCheckPath}
\title{Converts path to absolute path, checks to be sure it's an existing directory.
Optionally appends a fileName, if specified}
\usage{
dpCheckPath(path, fileName = "")
}
\description{
Converts path to absolute path, checks to be sure it's an existing directory.
Optionally appends a fileName, if specified
}

