% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpLoad.R
\name{dpCommitBatchData}
\alias{dpCommitBatchData}
\title{Commit (upload) data from a dpBatch to the associated database.}
\usage{
dpCommitBatchData(batch, fromDisk = T)
}
\arguments{
\item{batch}{a \code{\link{dpBatch}} object.}

\item{fromDisk}{when TRUE, commits the most recent copy of the batch, which
is loaded from the set folder stored on disk.  When FALSE, commits the
batch as it is passed to dpCommitBatchData. Tweaks can be saved to disk,
but not reflected in a batch resident in memory.  "fromDisk = T" as a
default is intended to ensure that any tweaks that were saved to the batch
on disk are included in the commit.}
}
\description{
Transfers the \code{data.frame} associated with a
  \code{\link{dpBatch}} object into the database described by
  \code{dpBatch}'s \code{\link{dpSet}}.  Any \code{\link{dpTweak}}s that are
  associated with the \code{dpBatch} are processed prior to loading the data
  into the database.
}

