% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R
\name{dpConnectionArgs}
\alias{dpConnectionArgs}
\title{Create a dpConnectionArgs object}
\usage{
dpConnectionArgs(drv, ...)
}
\arguments{
\item{drv}{DBIDriver object, name, or DBIConnection.  See \code{drv} argument
for \code{\link{dbConnect}} in the required \code{DBI} package.}

\item{...}{authorization arguments needed by the DBMS instance; these
typically include user, password, dbname, host, port, etc. For details see
the appropriate \code{DBIDriver}.}
}
\value{
A \code{dpConnectionArgs} object that can be passed to \code{dpSet()}
  or \code{dpConnect()}.
}
\description{
A dpConnectionArgs object is passed to the \code{dpSet()} function to create
a dpSet object. pdConnectionArgs objects specify the connect function (from
the DBI package) and the arguments passed to that function in order to create
to the set's database.
}

