% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpBatch.R, R/dpSet.R
\name{dpBatch}
\alias{dpBatch}
\alias{dpLoadSet}
\alias{dpSet}
\alias{is.dpBatch}
\alias{is.dpSet}
\title{Create and manage datapuppy \code{sets} and \code{batches}}
\usage{
dpBatch(batchRecord, set, dataValues, rawValues = data.frame())

is.dpBatch(x)

dpSet(setPath, connectionArgs, batchRowColumnName, datumValueColumnName,
  datumTypeColumnName, batchNameColumnName, batchesTableName, dataTableName,
  typesTableName)

dpLoadSet(setPath)

is.dpSet(x)
}
\arguments{
\item{batchRecord}{A named \code{list} of values describing the \code{batch}.
Values in the \code{batchRecord} are stored as a record in the batches
table. \code{Names} attribute of \code{batchRecord} must contain the name
of the database table column where each value is to be stored.}

\item{set}{A \code{\link{dpSet}} object describing the \code{set} to be
operated on (or the \code{set} to which a batch will be added).
Alternatively, a setPath (see 'setPath' argument in \code{\link{dpSet}})
from which a set will be loaded.}

\item{dataValues}{A \code{data.frame} with the values that are to be loaded
into the database.}

\item{x}{An object to be tested.}

\item{setPath}{A character string containing the path to the \code{set} (\code{batch})
directory.  This directory contains the \code{"dpSet.rData"} (]code{"dpBatch.rData}) file and a subfolder
for each \code{batch} included in the \code{set}.  The folder must exist and be
empty when a \code{set} is created using \code{dpSet()}.  If passed to
other functions as a 'set' argument, the folder specified by \code{setPath}
must contain a 'dpSet.rData' file.  If a path is not fully specified, it is
assumed to be a subdirectory of the R working directory (see
\code{\link{getwd}}).}

\item{batchRowColumnName}{A character string containing the name of a column
in the \code{dataTable} of the database; the column is used to store the
row number of the \code{dataValues data.frame} that was the source of the
datum.}

\item{batchNameColumnName}{A character string containing the name of a column
in the \code{batchTable} of the database that stores a unique name for the
batch.  This name should be meaningful to a human to identify the batch,
not the serial number (primary key) assigned to the batch automatically by
datapuppy.}

\item{batchesTableName}{A character string containing the name of the
\code{batchesTable} in the database.}

\item{dataTableName}{A character string containing the name of the
\code{dataTable} in the database.}

\item{typesTableName}{A character string containing the name of the
\code{typesTable} in the database.}

\item{batchPath}{same as setPath, above.}

\item{validate}{A boolean determining whether names in the argument list are
validated against \code{\link{formals}(dpSet)} or
\code{\link{formals}(dpBatch)}.}

\item{conndectionArgs}{A \code{dpConnectionArgs} object returned from
\code{\link{dpConnectionArgs()}} describing the location and credentials
for the database associated with a \code{set}.}
}
\value{
\code{dpBatch()} returns a \code{dpBatch} object and saves the
  arguments passed to \code{dpBatch()} in a subdirectory of the associated
  set's directory.  The subdirectory is named according to the value in
  batchRecord that has a name equal to set$db$batchNameColumnName.

\code{is.dpBatch()} returns TRUE if x is a \code{dpBatch} object.

\code{dpSet()} (\code{dpBatch()} returns a \code{dpSet}
  (\code{dpBatch}) object that describes the \code{set} (\code{batch}). This
  object should be assigned to a variable so that it can be passed to other
  datapuppy functions.  The arguments passed to \code{dpSet()}
  (\code{dpBatch()}) are also saved in a file on disk (see Details, above).

\code{dpLoadSet()} (\code{dpLoadBatch()}) creates a \code{dpSet}
(\code{dpBatch}) object described by the arguments stored in the
"dpSet.rData" ("dpBatch.rData") file.
}
\description{
Datapuppy \code{Batches} are collections of records derived from a single
data file (e.g., a data file downloaded from a data logger or a spreadsheet
that contains field observations).  Datapuppy \code{sets} are collections of
data \code{batches} to be loaded into a single database.
}
\details{
A \code{batch} is a collection of data points that have been imported from a
single data file.  For instance, when a data logger is downloaded, it creates
a data file with many records, where each record may contain observations of
several different metrics. All (or a subset) of those observations can be
collected into a \code{batch}.

A \code{set} is a collections of data \code{batches}.  A \code{set} is a
collection of all of the \code{batches} stored in a particular database.
Thus, each \code{set} is always associated with a single database.
\code{dpSets} and \code{dpBatches} are S3 objects built atop \code{lists}.
\code{dpSet} objects contains information about a \code{set} and the database
associate with the \code{set}.  \code{dpBatch} objects contain information
about the \code{batch} and the \code{set} to which the \code{batch} belongs.

Because \code{dpSets} (and \code{dpBatches}) are \code{lists}, information in a
\code{dpSet} (\code{dpBatch}) can be accessed with the $ operator.  For
instance, given a \code{set} named \code{mySet},
\code{mySet$db$keys$dataPrimaryKey} would return the name of the primary key
column for the data table in the database. Generally, tho, the user should
not have to investigate the contents of a \code{dpSet} (\code{dpBatch})
object. Instead, call \code{dpSet()} (\code{dpBatch}) to create the S3
object, store the object in a variable, and pass the variable to other
functions.

Note that datapuppy makes a few assumptions about the database into which
batches are loaded.  There must be a "batch" table where each record contains
infomation about a \code{batch} loaded into the database.  There must be a
"data" table where each record is a single datum.  And there must be a
"types" table where each record describes a metric that can be assoicated
with any datum (e.g., the datum represents a temperature reading, a wind
speed, a stock price, or whatever metrics are traked by the database). The
columns in the \code{batchesTable}, \code{dataTable}, and \code{typesTable}
of the database have some requirements:

1) Each table must contain an autonumber field that is designated as the
primary key for the table

2) the \code{dataTable} must contain at least two foreign keys, one that
refers to and is named the same as the primary key column of the
\code{batchesTable}, and one that refers to and is named the same as the
primary key column of the \code{typesTable}.  In this way, each datum is
associated with a \code{batch} that describes the source of the datum and
with a \code{datatype} that describes what the number represents.

\code{dpSet()} (\code{dpBatch()}) creates a file called
  "dpSet.rData" ("dpBatch.rData") in the location specified by \code{setPath}
  (\code{\link{file.path}(set$setPath, batchRecord$batchName)}). The file
  contains a list of the arguments passed to \code{dpSet()}
  (\code{dpBatch()}).  When a \code{set} (\code{batch}) is reloaded from disk
  using \code{dpLoadSet()} (code{dpLoadBatch()}), the list of the arguments
  is used to recreate the \code{dpSet} (\code{dpBatch}).
}

