% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlExample.R
\name{readICFNode}
\alias{createICFNode}
\alias{readICFNode}
\title{Input Control File Nodes}
\usage{
readICFNode(file)

createICFNode(path)
}
\arguments{
\item{file}{Name of the import control file to read.}

\item{path}{An optional root path that will be appended to datafile names in
the input control file.}
}
\value{
The contents of the import control file as an XMLNode object (see XML
  package documentation)
}
\description{
\code{Input control files (ICF) nodes} are \code{XML nodes} (see the
\code{XML} package documentation) that contain the information necessary for
\code{datapuppy} to create a database from raw data files.  The functions
documented here allow initial creation of an \code{ICF node} in memory, by
either reading a previously saved \code{ICF node} from a file or by creating
a blank \code{ICF node} template from scratch.
}

